requirements.txt
# Social Security AI Workflow Automation System
# Complete Python Dependencies for Production Deployment

# =========================================
# CORE WEB FRAMEWORK & API
# =========================================
fastapi==0.116.2
uvicorn==0.36.0
pydantic==2.11.9
pydantic-settings==2.10.1
python-multipart==0.0.20
starlette==0.48.0

# =========================================
# FRONTEND & UI
# =========================================
streamlit==1.38.0
streamlit_cookies_manager==0.2.0
plotly==5.17.0

# =========================================
# BACKGROUND PROCESSING & WORKERS
# =========================================
celery==5.3.4
redis==6.4.0

# =========================================
# DATABASE & ORM
# =========================================
sqlalchemy==2.0.43
alembic==1.13.0
psycopg2-binary==2.9.10

# =========================================
# AI/ML & DOCUMENT PROCESSING
# =========================================
ollama==0.1.7
qdrant-client==1.15.1
# easyocr==1.7.0  # Commented out due to Python 3.13 compatibility issues
transformers==4.35.0
torch==2.1.0

# =========================================
# WORKFLOW & STATE MANAGEMENT
# =========================================
langgraph==0.0.40
langchain==0.1.0

# =========================================
# FILE PROCESSING & COMPUTER VISION
# =========================================
PyMuPDF==1.26.4
Pillow==11.3.0
opencv-python==4.12.0.88
numpy==2.2.6

# =========================================
# AUTHENTICATION & SECURITY
# =========================================
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
email-validator==2.3.0
cryptography==46.0.1
bcrypt==4.3.0

# =========================================
# HTTP CLIENTS & API COMMUNICATION
# =========================================
httpx==0.28.1
requests==2.32.5
certifi==2025.8.3

# =========================================
# UTILITIES & LOGGING
# =========================================
structlog==25.4.0
typing-extensions==4.15.0

# =========================================
# TESTING FRAMEWORK
# =========================================
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-mock==3.12.0
locust==2.17.0

# =========================================
# DEVELOPMENT TOOLS
# =========================================
black==23.11.0
isort==5.12.0

# =========================================
# ADDITIONAL CORE DEPENDENCIES
# =========================================
click==8.3.0
h11==0.16.0
anyio==4.10.0
idna==3.10
sniffio==1.3.1
greenlet==3.2.4
annotated-types==0.7.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
dnspython==2.8.0
six==1.17.0
urllib3==2.5.0
charset-normalizer==3.4.3